separate main.cpp from multiblend
